// src/api/models/paymentListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Admin
 * Allta Admin 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { PaymentListItemUser } from './paymentListItemUser';
import type { PaymentListStoreItem } from './paymentListStoreItem';
import type { PaymentListPurchaseItem } from './paymentListPurchaseItem';

export interface PaymentListItem {
  /** ID */
  id: string;
  /** 결제 상태 */
  status: string;
  /** 상품 종류 */
  productType: string;
  /** 서비스 종류 */
  serviceType: string;
  /** 매장 이름 */
  storeName: string;
  /** 변동 금액 */
  amount: number;
  /** 생성일 */
  createdAt: string;
  /**
   * 결제 회원 정보
   * @nullable
   */
  user: PaymentListItemUser;
  /** 결제 매장 정보 */
  store: PaymentListStoreItem;
  /** 구매 정보 */
  purchase: PaymentListPurchaseItem;
}